Reorganize build trees for generated libs
So far, the build tree for generated libs had a layout that was
supposed to mimic the installation tree. This immitation has always been
imperfect: for instance, installed shared libs are supposed to be in
$PREFIX/lib on Unix and $PREFIX/bin on Windows, and they were always put
in $BUILD/lib in build trees. That specific layout introduced
complexity for little gain.

This commit reorganizes build trees to be more dev-friendly:

* $BUILD/*.gpr: project files (generated library and mains);

* $BUILD/src: generated sources for the library; keeping static files in
  their original location allows incremental rebuilds without
  re-generating the lib, which is dev-friendly;

* $BUILD/lib and $BUILD/obj: library and object directory for the
  generated lib;

* $BUILD/src-mains: generated sources for the mains;

* $BUILD/obj-mains: object directory (and exec directory) for the mains;

* $BUILD/scripts: generated scripts (for instance, the playground);

* $BUILD/gdbinit.py: generated initialization script for GDB helpers.

Directories for language bindings ($BUILD/python and $BUILD/ocaml) stay
as they are today. As a result, everything that developpers interact
with directly is either in the build directory or one of its
subdirectories, and all paths in project files are simpler (no more
"../../").

Note that, due to a bug in gprinstall (TA20-017), we cannot put the
C header for the generated library ($lang.h) in the project's sources
($BUILD/src). Instead, for the time being we will put it directly in
$BUILD, install it manually and add that directory to C_INCLUDE_PATH.

TN: T914-010