Fix lang_source_dir for testcases
So far, ManageScript assumed that the lang_source_dir is the directory
that contains the Python source file which defines the ManageScript
subclass. This is wrong for testcases, as there is generally only one
ManageScript subclass, in the testsute/python_support/utils.py source
file.

Introduce a new "root_dir" ManageScript constructor argument to override
this behavior, and use it in the testsuite appropriately.

TN: T914-010