langkit.lexer.Lexer: fix the "signature" property
This property never worked, in a disturbing way: it has always raised an
AttributeError for every call. Because of arcane Python semantics, this
used to redirect the evaluation of this property to fallback on the
__getattr__ method, which just looked up the "signature" property on the
LexerToken subclass.

As a result, the effective signature of the whole lexer used to be
restricted to information on the LexerToken subclass, thus preventing
the generation of lexer code for a whole class of lexer changes in the
DSL.

TN: TA28-011