DynamicLexicalEnv: add an entity resolver argument
This commit adds an optional resolver argument to DynamicLexicalEnv
constructions in the DSL. This resolver works in a similar way to
add_to_env's "resolver" argument: it is called to refine the result of
lexical env lookups.

For clarity, this renames DynamicLexicalEnv's current "resolve" argument
to "assocs_getter", and the new argument is called "assoc_resolver".

TN: TC17-036