AbstractNodeData: make it possible to override access construction
This commit makes it possible to customize the construction of
FieldAccess expression for specific fields/properties:
FieldAccess.construct can now use a hook to build the resolved
expression that implements the field access/property call in the DSL.

An upcoming change will introduce iterator types, and in particular a
"to_iterator" property that takes no argument at the DSL level, but
takes an analysis context in the generated code. This overriding
mechanism will enable to implement the compilation of calls to this
property.

TN: T218-020