Introduce error nodes for Skip parsers
This commit first introduces a new annotation for nodes: "error_nodes",
with a set of restrictions. It then updates Skip parsers to only allow
the creation of error nodes, so that the restrictions also apply to the
nodes that Skip parsers create.

This opens the way to fix unparsing unsoundness for such nodes, which
cannot be unparsed: Skip parsers do not add syntax information to the
nodes they create.

TN: T520-011