PropertyDef: introduce the "artificial" constructor argument
A future change will introduce internal properties for which we always
instantiate PropertyDef in Langkit, but that make it to the generated
Ada code iff they are reachable (i.e. transitively used from a public
API).

Given the state of Langkit so far, this means either:

* Making these properties not internal, but this will create lots of
  unused properties (users will be warned about them).

* Making these properties internal, but this means that our reachability
  analysis will consider them as used, and thus we will end up
  generating Ada code for all of them.

This commit introduces a new "artificial" attribute to properly describe
these new properties:

* Language specs can refer to these artificial properties (unlike
  internal properties).

* Langkit will never warn about unused artificial properties (unlike
  non-internal properties).

TN: T218-020