$.Lexical_Envs_Impl: use rebindings version info to check if stale
Thanks to changes on T121-033, rebindings have their own version number
that we can use to check if rebindings are stale. Replace custom and
hackish use of context version with this. As a nice side effect, this
also fixes the rebinding of the empty environment, which used to crash
with a Constraint_Error, as the empty environment has no owner.

TN: U219-034