Introduce the file reader abstraction
This commit introduces a new abstraction (File_Reader_Interface in Ada)
so that users of the generated libraries can override how the contents
of source files are fetched. This opens the way to integrated
preprocessors, but also makes it easier for projects such as IDEs to
intercept the reading of source files to provide in-memory buffers for
files that open for editing.

TN: U310-050