$.Introspection: fix order in node derivations and First_Kind/Last_Kind
A missing sort for node subclasses in ASTNodeType.subclass used to put
node derivations in the wrong order in the introspection API, which gave
invalid input to the First_Kind/Last_Kind functions leading to wrong
results. This commit adds the required sorting, which fixes all these
bugs.

TN: U412-016