Langkit_Support.File_Readers: file reading errors in Direct_Read
This commit adds a return statement in Direct_Read so that, in case of
Name_Error exception (the file could not be open), Direct_Read does not
continue using GNATCOLL.Mmap on a not-open file (it used to raise
Constraint_Error exceptions in such cases).

This also extends testing to cover such error cases.

TN: U422-030