Make it possible to bypass file readers to parse internal units
It is often necessary for generated libraries to handle some "intenal"
unit, that will define language builtins. This unit is generally not
represented as an actual source file ("filesystem" tense), but rather as
an in-memory source buffer, which is incompatible with file readers,
which currently forbid all in-memory parsing.

This commit introduces an internal hook (a new "Is_Internal" formal in
$.Implementation's Get_Unit) so that generated libraries can bypass file
readers and thus load their internal units from memory at all times.

TN: U422-030