Remove recursive modules from OCaml bindings
To be able to remove recursive modules, we choose to compile the ocaml
bindings slightly differently. The public API remains almost untouched.

Before the strategy was to have recursive modules containing the
definition of the ast node type together with its wrap function. Now we
first emit the recursive types at top level. And then we emit the
mutually recursive wrap functions.

We still have another set of modules representing the ast node types
together with the functions and properties associated with this node.
This second set of modules does not have to be recursive since its using
the types already defined.

TN: T812-013