Lexical envs: cache empty envs in dynamic env getters
This commit first brings context version info to lexical envs internals
in Langkit_Support and then uses it to properly cache (i.e. invalidate
when needed) empty environments in dynamic env getters.

Caching these environments is necessary to get reasonable complexity
when we fail to resolve referenced environments: in such cases we return
the empty env, and we need to cache it in order not to spend extra time
re-resolving it in the future, which can happen very often when multiple
referenced environments are involved.

TN: U512-020