Update langkit tutorial
This patchs updates the langkit tutorial, starting from the `Getting
Started` section. It consists of updating the new file/path names used
by langkit when creating a project and switching some Python code from
version 2 to 3.

It also adds a new file, `kalint.py` to help the reader to correctly
assemble all the pieces of the langit toy interpreter.