U916-032: declare unit Node method to LKT prelude
Adds unit() method in Node class to LKT prelude. It returns an
AnalysisUnit, so this patchs also declare this new class.

The following tests are now passing lkt_semantic_checks:
- testsuite/tests/properties/analysis_unit/
- testsuite/tests/misc/event_handler/

This one pass the sematic checks but not the legality one:
- testsuite/tests/properties/memoized_unit_loading/

This change also impacts the following tests, but some builtins are
still missing to make them pass:
- testsuite/tests/properties/analysis_unit_from_node/
- testsuite/tests/properties/null_checks/
- testsuite/tests/misc/event_handler/
- testsuite/tests/properties/unique/
- testsuite/tests/properties/then_analysis_unit/
- testsuite/tests/properties/invalidation_during_mmz/
- testsuite/tests/lexical_envs/ple_after_reparse/
- testsuite/tests/misc/event_handler/