lkt/language/parser.py: raise an exception if Node has no type
We're using cast_or_raise here is because this is used in internal
methods where we expect to return a type. Here, it made the node_type
property return None, which is clearly not great, rather than raising
an exception, which is better because it announces that underlying
assumptions have been broken.