U930-039: Lkt: support ArrayLiteral type resolution
ArrayLiteral expected_type resolution was missing, leading to spurious
failures. Therefore, this patch fixes that and implements the typing
of array literals: a) by using type information coming from the
context, or b) by infering the type from the context-free type of the
array elements.

For example, this patch allows to infer ArrayLiteral type, as well as
its elements type, from the elements context-free type:

```
fun f(): Int = 1
val a = [0, f(), 2]
```

Type of `a` is now inferred from the type of `f()` as soon as all the
elements have the same, or compatible, type. Therefore, type of `0`,
and `2` can also be inferred from the freshly inferred array's type.