U930-039: fix Lkt's ArrayLiteral expected_type resolution
ArrayLiteral expected_type resolution was missing, leading to spurious
failures. As a first step, this patch implements the typing of array
literals when the information for element type comes from the context.
Nevertheless, the parser is not able yet to infer the type of array
literals from the context-free type of their elements.