Python API: fix unwrapping of arrays of structs
Unwrapping an array of structs is meant to take the ownership share for
these structs, and thus to clear them. This means that the destructor of
Python wrappers for structs must handle the case where they no longer
have the ownership of a struct value: they didn't so far, which
triggered crashes.

TN: UA05-027