Replace "string_repr" with "ascii_repr"
The "langkit.common.string_repr" function has never been complete (it
could emit an invalid Ada literal for an otherwise valid string, for
instance for linefeed characters). Moreover, since the transition to
Python3, it used to accept Unicode strings whereas it is generally used
to emit literals for Ada's String type (which contains bytes).

This commit renames it to "ascii_repr", to make it clear that is expects
only ASCII codepoints in the Unicode string it takes and makes it use
"bytes_repr" to make the implementation correct for special characters.

TN: UA05-027