Lkt: improve raise expression type resolution
As the issue reported in U928-029 cannot be reproduced, add support
for the PropertyError object type resolution in order to pass the
semantic checks (for memoized_property_error).

Add support in ``expected_type`` in order to properly getting type
from a ``BlockExpr``. This patch also improves the support of the
raise expression as it only accepts exception expressions
(``PropertyError``) now.

Moreover, this change allows to enable semantic checks on
memoized_unit_loading, and let_underscore tests.

TN: U928-029