langkit.emitter: re-add bodies in the library interface
If "libadalang-analysis.ads" is in the library interface, but not
"libadalang-analysis.adb", GPRinstall does not install
"libadalang-analysis.ali", so add bodies back, at least to avoid a
general build failure tomorrow.

Ahead: determine whether this is acceptable, and fix GPRinstall if it is
not (we will revert this patch in that case).

TN: UA21-007