Langkit_Support.File_Readers: fix invalid memory access
Recent changes in this unit introduced an invalid memory access
(which Valgrind finds as a conditional jump depending on uninitialized
value): when no codepoint could be decoded from the source buffer,
the Output_Index and thus Contents.Last indices point to uninitialized
source buffer locations, so we really should not try to read the source
buffer to compute location. It is not needed anyway: this is the
specific case of the first codepoint in the source buffer, so we know
its location: line 1, column line 1.

TN: UB12-067