Lkt: merge LktNode.check_legality and LktNode.check_semantic
This commit unifies code that does legality check and code that
processes semantic results, so that errors are the same in both cases.
(we used to detect error message-less erroneous semantic results in the
former only)

It also fixes TreeSemanticResult.has_error computation in check_semantic
to set it to true iff there is an error diagnostic in the result.

TN: RA22-015