Symbolic solver: remove support for Comparer_Type
The use of equality predicate is unsound. For instance, given the
Modulo_Comparer comparator, which considers that X and Y are equivalent
iff X mod 10 = Y mod 10, then the solution of the following equation
depends on the evaluation order:

   Assign(X, 1)
   and Assign(X, 11, Modulo_Comparer)

Get rid of this feature, which simplifies and removes other unsoundness
from the symbolic solver (Can_Fail currently defeats the dataflow
analysis).

TN: SB20-024