Symbolic solver: make it possible to evaluate partial sequence of atoms
This makes it possible to evaluate a sequence of atoms disregarding the
atoms using variables that no atom defines (i.e. partial sequences of
atoms). Future commits will use this capability to look for
contradictions in partial developments of relations which in turn will
be useful to implement optimizations.

TN: SB20-024