Symbolic solver: add an optimization to split Any relations
This new optimization is run before running the symbolic solver itself,
to reduce the number of disjunctions in the relation to solve, as these
are responsible for the solver's algorithmic complexity.

The way it works it to perform a depth first traversal on the relation
tree, collecting atoms along the way and detecting contradictions
between these atoms: when a contradiction happens, the currently
processed Any relation alternative can be removed.

TN: SB20-024