Rework adalog tests
Move as much code as possible from the generation scirpt to
Langkit_Support (modifying code templates is always a pain). Also make
it possible for tests to restrict the set of solver configuratins to
exercize:

* the default (symbolic without optimizations, symbolic with
  optimizations and state machine);
* only symbolic without optimizations;
* only symbolic with optimizations;
* only state machine.

Some upcoming tests will check optimizations with relations that are too
complex to solve with plain solution exploration in reasonable time, so
restricting runs to optimizations is necessary.

TN: SB20-024