Symbolic solver: fix Simplify crash in debug mode
When Simplify removes an All alternative in an Any relation, it should
stop iterating on sub-relations of the All relation: they are no longer
part of the relation to simplify. This used to trigger a crash, as the
local variable for the Any relation was set to null in this case,
and debug code tries to use it.

TN: SB20-024