Remove useless recursions in parsers handling
The only thing needed to generate the code for a Defer parser is the
name of the function that implements the deferred parser. It is thus
completely useless to trigger the generation of the function body for
that function at this point.

This commit just stops doing that (the code for all parsers is still
generated with a GrammarRulePass anyway), which removes this "chained
parser code generation" behavior. This fixes the Python stack overflow
that could come from it.

This commit also replaces a completely recursive parsers traversal
starting from the main rule with a stack: the traversal of parsers for a
given grammar rule is still recursive, but at least traversing a parser
cannot trigger the traversal of another grammar rule's parsers anymore,
which fixes another risk of stack overflow.

TN: V103-031