langkit.diagnostics: remove mandatory import of Mako for GDB helpers
A previous commit added in langkit.names an import of
langkit.diagnostics. Since langkit.diagnostics imports
langkit.documentation, which eventually imports Mako, this meant that
importing langkit.names required to import Mako, which is not acceptable
for GDB helpers (which need to work with GDB's standard Python
distribution).

TN: V126-009