Generic API: fix the handling of overriding properties
The Langkit_Support.Generic_API.Introspection.Members function used to
crash when called on nodes which define overriding properties. The
problem was that, in introspection tables for such nodes, the number
of inherited members was under-estimated.

That was due to a bad interaction between CompiledType.get_properties,
which always return the "most derived" property on a give node, and the
properties dispatch lowering pass, which makes all properties private in
a derivation tree. As a result, CompiledType.get_properties could return
only one private property, hiding the public root one.

This is fixed by getting dispatcher properties before checking whether
they are public when generating introspection tables.

TN: U401-024