Properties: fix code generation for token arrays
When properties logging is enabled and one property refers to the "token
array" type, we generate a `Trace_Image` function overload for token
arrays, which expects a `Trace_Image` overload for tokens to exist. It
did not so far (an oversight), so add one.

Found while writing a testcase for GDB helpers.

TN: S521-024