Adalog: remove the Var_Or_Null type
In practice, logic variables are pointers, and thus are nullable types,
so there is no need for Var_Or_Null, which is an "option" wrapper:
either we have a null Logic_Var (i.e. no variable), either we have a
non-null one.

This is a very minor optimization, as the dependency analysis primitives
now only accept/return a pointer instead of a pointer + a boolean.

TN: SB20-024
For libadalang/langkit#618