$.Implementation: remove the protected object and use GNAT.Task_Lock
Using a protected object pulls libgnarl, which is undesirable for the
users of some Langkit-generated libraries. GNAT.Task_Lock is a good
enough replacement for that: it does not pull libgnarl, and the fact
that this lock is global is not a problem as it is re-entrant and used
in a context that is not performance sensitive (analysis context
allocation/dellocation).

TN: U907-017