pkg_unparsing_impl_body_ada.mako: do not use Child function for RW nodes
An upcoming change will remove indexes from the rewriting API, so this
function will no longer be available.

Note that eventually, the representation of list nodes will switch from
children vectors to doubly linked lists: when that is the case, we will
need to introduce an iterator type for abstract nodes. We will cross
that road when we will get there.

TN: RC13-032