GenericAPI: remove the "Entity_" prefix from short type names
We never need to distinguish between node types and the corresponding
entity types in this context (this is because only entity types are
exposed). To simplify the generated code, drop the "Entity_" prefix for
the short type name for entities (just like we do in public APIs). This
has no visible effect in the public API, so no test updated.

TN: RC13-032