Lkt: fix wrong chaining in Expr.scope
The following expression:

  X.then(Y)._or(Z)

will compute Z if X is true while Y is false.

In Expr.scope this is not what we want to do, as we check several cases
that are supposed to be mutually exclusive: if we are processing a
DotExpr, we only want to compute the designated scope from the prefix:
we do not want to fallback on the "regular environment" case when the
prefix is erroneous.

This commit fixes this, which prevents Expr.scope to return a non-empty
environment for a DotExpr when there is an error analyzing the prefix.
This in turn avoids the spurious "cannot find entity `Y` in this scope"
error for "X.Y" when there is already a "cannot find entity `X` in this
scope" error.

For #622