Improve the error message for parserless nodes
The current message is incomplete:

    This node is not synthetic nor abstract, so at least one parser must
    create it

Since enum node alternatives currently dont have any location
information attached to them in the Python DSL, and that their location
in Lkt is the "parent" enum class, this error message does say which
alternative has no parser. This commit adds the DSL name of the
parser-less node, which is much more helpful.

For #628