Lkt lowering: fix lowering order for list node derivation
Depending on the order picked by Langkit for node type lowering,
processing the definition of a derived list node could crash. The
problem was that lowering code just assumed so far that the root node
was lowered before list nodes, and that list element nodes were lowered
before the corresponding list nodes as well, whereas this has never been
guaranteed.

To fix this, we now force the lowering of the element node when lowering
a list node, and stop assuming that the root node is lowered first.

For #622