Generic API: make "=" work on non-class-wide Lk_Node
  function "=" (Left, Right : Lk_Node'Class) return Boolean;

is not picked up by generic instantiations such as:

  package Lk_Node_Vectors is new Ada.Containers.Vectors
    (Positive, Lk_Node);

as Ada's rules mandate that the package instantiation takes Lk_Node
built-in "=" operator instead.

Change Langkit_Support.Generic_API.Analysis."=" to work on
non-class-wide Lk_Node types instead, so that it is used instead.
Lk_Node is not supposed to be extended ("type X is new Lk_Node with
..."), so this should not break any existing code.

TN: U607-013