ArrayType.builtin_properties: set lazy_field to False
Whether properties are lazy fields is determined in the "compute
property attributes" compilation pass. However, array types, and thus
their "to_iterator" properties, can be created later on (when
constructing property expressions). As a result, in order for the "lazy
field" information to be properly initialized for "to_iterator", we need
to fill it in right at property creation.

For #630