Properties: add "Self" sloc to messages of implicit Property_Error exc.
When a property raises an error, for instance because it tries to get
some field in a null node, it would be nice to include the sloc of the
node in that property, so that error messages give users a clue about
how it relates to the Ada source code that Libadalang analyzes, which
would be useful to workaround or investigate an issue.

Note that adding this sloc information is done on a best effort basis.
As a notable example, the Property_Error exception raised in the lexical
environment lookup system do not provide it, as the "Self" node for the
property that triggers a lexical env lookup is not available there, and
taking this information to this point would require a lot of logistics,
and possibly run time performance degradation.

Closes #635