Properties: preserve the exception message for memoized properties
When called for the first time, a memoized property that raises an
exception lets the exception propagate with an original exception
message. However, the next times it is called, the exception that is
propagated has just “Memoized error” as its exception message.

This commits fixes this so that the exception that is raised preserves
the same original message. It also includes "(memoized)" to help
diagnosing semantic analysis issues in production. This is a minor
breach to the principle that memoized property cache hitting calls must
behave like cache missing ones, but should be of no consequence wrt
soundness: properties catching exceptions should not rely on the
exception messages, which are only lossy debugging aids.

Closes #636