Properties: do not overwrite cache entries for memoized exceptions
This is a bug in the sense: not working as expected, but producing
exactly the same results. When evaluating a memoized property that has
raised an exception, memoization code re-raises the exception (as
expected), but then this triggers the exception handler that keeps track
of the fact that the memoized current property raised an exception.

In other words, subsequent calls to a property that raises an exception
write again and again an entry in the memoization table that is
identical to the entry that was already there. This commit re-organizes
memoization code to avoid this. No change visible through the API, so no
testcase added...

For #636