Adjustments to the GitLab CI
Add workflow rules to prevent duplicated runs between branch pipelines
and merge request pipelines.

Add an omitted "anod vcs" in the test job.

Trigger the Libadalang pipeline as part of the test stage.

Upload the package as part of the test stage, so it's done systematically
in reaction to a push on main, even if the tests fail.