Fix possible dangling references
Ctypes struct are more or less pointers to real C structures. Thus,
pointers contained inside struct are not managed by the OCaml value
representing the struct. In case a pointer allocated with Ctypes.allocate(_n)
is assigned to a field of a struct, the actual ocaml value containing
the result of allocate can be GCed and the underlying pointer freed,
which means that the field contains a dangling pointer.