.gitlab-ci.yml: fix testing
Now that access to Cathod is enabled, Anod no longer plans to build
langkit and langkit_support: it tries to install them. Update the
"assume-untouched" hacks to cover the installation instead of the
builds, so that we use the langkit/langkit_support that were just built
instead of the wavefronts during testing.