$.Implementation: add a cache around unit providers
The recent extension of the unit provider interface created necessary
performance regressions in Libadalang. It is caused by both more work in
the unit providers (they compute/return more information) and because we
decided not to break the existing unit provider API: it is not possible
to use more efficient data structures.

To reduce the regression and get even more performance gains, introduce
a cache so that unit providers are used as little as possible.