Fix and harmonize the handling of external variables in project files
For some settings, like the build mode or the library kind, the current
community-friendly scheme we want is to have both:

* an external variable (`MYVAR`) that is able to set the setting for all
  projects that support it and

* an external variable (`MYLIB_MYVAR`) to change the setting for only
  one project.

The correct GPR incantation for this is:

```
My_Var := external ("MYLIB_MYVAR", external ("MYVAR", "default"));
```

So that when setting both `MYLIB_MYVAR` and `MYVAR`, it's the
project-specific one (`MYLIB_MYVAR`) that takes precedence. We used an
incorrect incantation so far:

```
My_Var := external ("MYVAR", external ("MYLIB_MYVAR", "default"));
```

Fix this for all relevant external variables in `langkit_support.gpr`
and Langkit-generated project files, and harmonize GPR code style for
the other variables.