C API: remove kind/text/sloc range from the token data type
This type is meant to designate a token reference, so it makes more
sense to have it contain only data that allows to reach storage for the
rest of the token data: token data handler pointer and token indexes.
Remove the kind/text/sloc range data from this type and add token
primitives to fetch them instead.

Update Python, OCaml and Java bindings accordingly.

Note: this is preparatory work for the switch of source buffers to
UTF-8. It is needed in order to avoid dynamically allocated resources
(text objects) inlined in structs (token dta types) in the C API, which
is delicate to handle in language bindings. Removing text objects from
these structs and instead add accessor functions is much more practical.