langkit.compiled_types: compute inheritance info for all abst. node data
So far, this information was computed 1) for properties and partially
for syntax fields and 2) during some passes, so not always available.

This change extends `AbstractNodeData` and `CompiledType.add_field` so
that this information is computed just after fields are created and
for all kinds of fields (properties, syntax fields, user fields, builtin
fields). It also removes all the now-redundant bits for equivalent
features from `Field` and `PropertyDef`, removes obsolete hacks and
finally adjusts testcases to expected behavior changes.