Introduce PropertyDef.is_dispatching_root
After the property dispatcher lowering pass has run, the inheritance
information for properties is corrupted: the properties' `.base`
property is not `None` whereas no property is supposed to be deriving
from each other at this point. The `dsl_unparse` pass relies on this,
and will break once the inheritance information is fixed by upcoming
changes. Introduce a new flag for properties so that `dsl_unparse` no
longer relies on this data corruption.