Java: fix AnalysisContext.eventHandler initialization
Through the lifetime of a process, the same C-level physical object for
an analysis context may be reused for multiple logical contexts. As a
result, in the Java API we may have a single AnalysisContext instance
used multiple times as well, but each time with a different event
handler. As a consequence, the AnalysisContext.eventHandler field must
be updated each time we reuse the AnalysisContext instance.

This change makes the AnalysisContext.eventHandler private field mutable
and adds a fromReference method overload to update it. This overload
will enable language specs extensions which introduce anlysis
context-creating functions to correctly initialize this field.