langkit.reset: reset tokens
Parts of the generated code assume that the representation of the token
kind enumeration is "natural": first value is represented by 0, second
by 1, etc. This assumption was violated for tests ran with
`types_from_lkt=True`: the lexer was compiled a first time (to unparse
the DSL), and then a second time (to compile from Lkt), so tokens were
registered twice, creating gaps in assigned enum representation codes.
Extend the existing mechanism to "reset" langkit to cover these codes,
which restores the broken assumption.