Lkt: fix the parsing of logical expressions
Before this patch, the `%foo().solve()` expression was parsed as:

    LogicExpr(
      expr=CallExpr(
        name=DotExpr(
          prefix=CallExpr(name=foo, args=[]),
          suffix="solve",
        )
        args=[],
      )
    )

Which is unfortunate: the logic expression is meant to express the
construction of a logic equation. This patch modifies the parser so that
the parse tree for the above expression becomes:

    CallExpr(
      name=DotExpr(
        prefix=LogicExpr(
          CallExpr(name=foo, args=[]),
        ),
        suffix="solve",
      )
      args=[],
    )

i.e. the call to "solve" is done on the logic equation itself
(LogicExpr) rather than on the call to "foo" directly.