langkit.reset: clear memoization caches first
Doing this first is necessary for builtin types to be actually
re-created: upcoming work will add missing `T.env_assoc` and
`T.inner_env_assoc` creation in
`langkit.compiled_types.create_builtin_types()`. These are implemented
with memoized functions, so in order for the creation to really happen,
the memoization tables needs to be cleared first.