Logic predicates: fix the handling of default values + dynamic vars
The logic to turn partial arguments into list of arguments in a property
call was incorrect: when there are natural arguments with default values
in addition to dynamic variables, emitting positional arguments
accidentally associated dynamic variable to the arguments with default
values. For instance, for the property:

    @with_dynvars(my_var)
    fun pred(n: Entity[FooNode], flag: Bool = false)

then, creating the following logic predicate:

    bind my_var = 0;
    %predicate(FooNode.pred, node.lvar1, node.lvar2)

used to generate the following property call:

    Foo_Node_P_Pred
      (Node,          --  For "node"
       Entities (1),  --  For "n"
       0);            --  For "my_var"

whereas the property signature is actually:

    Foo_Node_P_Pred
      (Node: Foo_Node, N: FooNode, Flag: Boolean, My_Var: Integer);

This change refactors the property partial evaluation handling for logic
predicates to keep track of which argument/dynamic variable is passed,
and to emit property calling code correctly.