C API: fix type names for arrays
In the C API, the current name for arrays of symbols is
`unbounded_text_type_array`, which is confusing, especially given that
it implements an array of `symbol_type` values encoded as symbols.
Adjust names computation so that it is called instead
`symbol_type_array`. Note that this changes logic that applies to all
array types, so all array types benefit from similar enhancements.