C API: define the default grammar rule as a macro
Defining it as a "const" declaration in a header file is incorrect, as
it defines a symbol: two compilation units that include the header file
will then trigger link error due to conflicting symbol definitions.

Define it as a macro instead to avoid this kind of issue. No new
dedicated testcase for this as the next commit reworks C test helpers to
have multiple compilation units (several "*.c" sources in test
projects).