Unparsers: always include text in token unparsers
So far, when processing a token for which the lexer assigns always the
same literal, unparsing tables omitted the token text, and so unparsing
code had to use language-specific functions to get the text to unparse,
making it impossible to implement the same logic in Langkit_Support
(i.e. without language-specific functions available).

Always include token texts in unparsing tables to fix this problem.