setup.py: rename Windows executable from 'create-project.py.exe' to 'create-project.exe'
Setuptools generates an executable wrapper file for script entry points
which is not a Python script per se. For example, on Windows this is
a standard PE binary which invokes the Python interpreter with the
configured script.

Therefore, it is not necessary to add the '.py' suffix to the entrypoint
script name.